When connected to a Tor proxy, only use onion-servers
